Extract constructor and bootstrap methods into separate files
