+++ packages/truffle-contract/lib/contractHelpers/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  constructorMethods: require("./constructorMethods")
+};
